Move stream shutdown signal handler to _stream_hub
The signal handler was only in _flask.py, so FastAPI and Quart never
got it. The lifespan shutdown hook fires too late (after connections
drain), but connections can't drain because the streaming generator
blocks them. Move the handler to _stream_hub.py which all backends
import, so SIGINT tears down streams before the server starts waiting.
